perf: load the anonymous Apex tool when it runs
The tool calls `@salesforce/core`, which costs about 250 ms to load, and a
session that only reads logs must never pay it. `src/server.ts` now refuses a
disabled call, then loads the tool with `await import()` inside the handler.

`src/tools/executeAnonymousDefinition.ts` holds what `tools/list` puts on the
wire, so registration stays synchronous and the served definition does not
change. The refusal moves to `src/policy/orgExecutionPolicy.ts`, which already
owns its message, so a server started with `--no-apex-execution` loads no SDK
at all.

Starting the server falls from about 290 ms to about 55 ms, and loads none of
the SDK where it loaded 83 files of it.

Two guards keep it that way, and both fail on the parent commit: an ESLint
rule over `src/`, and a test that mocks the package with a factory that
throws. The jest transform gains `verbatimModuleSyntax`, as tsconfig.json sets
it, because swc otherwise drops a type-only import and the test cannot see
what the built output loads.

Refs #165
