+++ CLAUDE.md
@@ -105,4 +105,6 @@ Anonymous execution (4) accepts multi-line strings containing Apex, saves the re
 
 `apexlog_execute_anonymous` is **always registered** so agents can discover it; authorization happens per call in `src/policy/orgExecutionPolicy.ts`. `src/salesforce/orgClassification.ts` identifies the target org (`sandbox`, `scratch`, `trial`, `developer`, `production`, or `unknown` when it cannot be queried), caching one `Organization` query per org id for the server's lifetime. Non-production orgs run silently. Production and `unknown` orgs need either the `--allow-production-orgs` flag or a per-call user confirmation via MCP elicitation; without either, the call is refused with an `isError` result whose text explains both routes. Authorization runs before any `DebugLevel` or `TraceFlag` record is created, so refused calls leave the org untouched. `--no-apex-execution` refuses every call before contacting Salesforce and marks the tool `[DISABLED on this server]` in its description. The 1.x `--allowed-orgs` flag is accepted but ignored, with a stderr deprecation warning.
 
+`@salesforce/core` is a value import only in the two modules that call it, and `src/server.ts` loads `executeAnonymous` with `await import()` inside the handler — its wire definition lives in `executeAnonymousDefinition.ts`, so registration stays synchronous. A session that only reads logs never loads the SDK: startup is about 55 ms, not 290 ms.
+
 `src/index.ts` stays the `bin` entry point (`dist/index.js`) and does nothing but parse flags and call `runStdioServer`. `src/server.ts` holds `createApexLogServer`, `runStdioServer` and `parseServerConfig` and is free of import side effects, so tests can import it without spawning a server or parsing the test runner's own argv.
