+++ jest.config.js
@@ -13,6 +13,11 @@ export default {
         jsc: {
           parser: { syntax: "typescript", decorators: true },
           target: "es2022",
+          // As `tsconfig.json` sets it. Without it swc drops an import whose
+          // names are all used as types, and a test cannot see what the built
+          // output loads — which is what tests/salesforceCoreIsLazy.test.ts
+          // asserts.
+          transform: { verbatimModuleSyntax: true },
         },
         module: { type: "commonjs" },
       },
