+++ src/policy/orgExecutionPolicy.ts
@@ -90,6 +90,27 @@ export const APEX_EXECUTION_DISABLED_MESSAGE =
   "Anonymous Apex execution is disabled by server configuration (--no-apex-execution). " +
   "The log analysis tools remain available.";
 
+export function toolError(text: string) {
+  return {
+    content: [{ type: "text" as const, text }],
+    isError: true,
+  };
+}
+
+/**
+ * The refusal a server started with `--no-apex-execution` owes every call, or
+ * `undefined` when execution is allowed.
+ *
+ * Asked by the registered handler before it loads the tool, and by the tool
+ * itself for a direct caller, so the decision exists once and a disabled server
+ * loads no Salesforce SDK at all.
+ */
+export function apexExecutionRefusal(apexExecutionDisabled: boolean) {
+  return apexExecutionDisabled
+    ? toolError(APEX_EXECUTION_DISABLED_MESSAGE)
+    : undefined;
+}
+
 function truncate(apex: string, max: number): string {
   return apex.length <= max ? apex : `${apex.slice(0, max)}\n... (truncated)`;
 }
